refactor: restructure legacy app-view
